+++ templates/python.restful/{{cookiecutter.project_safe_name}}/app/__init__.py
@@ -14,7 +14,7 @@ cors = CORS()
 
 
 def create_app(config_name):
-    app = Flask(__name__)
+    app = Flask(__name__, static_url_path='/{{cookiecutter.project_slug}}/static')
     app.config.from_object('config.default')
     app.config.from_object('config.{0}'.format(config_name.lower()))
     app.config['VERSION'] = __version__